+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Wrap text in form elements.
 - Trigger an event on gl-like-button click.
 - Wrap basemap names in switcher.
+- Added header facet widget.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.